+++ src/extension.js
@@ -737,10 +737,14 @@ function activate( context )
 
     function refreshFile( document )
     {
-        function addResult( offset, removeLeadingComments )
+        function addResult( offset, removeLeadingComments, maxLength )
         {
             var position = document.positionAt( offset );
             var line = document.lineAt( position.line ).text;
+            if( maxLength !== undefined )
+            {
+                line = line.substr( 0, maxLength );
+            }
             if( removeLeadingComments === true )
             {
                 line = utils.removeLineComments( line, document.fileName );
@@ -763,7 +767,7 @@ function activate( context )
             {
                 var extractExtraLines = function( section )
                 {
-                    result.extraLines.push( addResult( offset, true ) );
+                    result.extraLines.push( addResult( offset, true, section.length ) );
                     offset += section.length + 1;
                 };
                 var isMatch = function( s )
@@ -789,8 +793,7 @@ function activate( context )
                     var offset = match.index;
                     var sections = match[ 0 ].split( "\n" );
 
-                    var result = addResult( offset, false );
-
+                    var result = addResult( offset, false, match[ 0 ].length );
                     if( sections.length > 1 )
                     {
                         result.extraLines = [];