+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improve colour validation for icons
 - Fix theme colours for ruler colour
 - Remove trailing block comments when adding multiline todo to tree
+- Support multi-line haskell comments
 
 ## v0.0.206 - 2021-03-17